fix (readme): remove settings_file config info
Settings_file is a cli command line keyword argument (which can specify
the location of zappa_settings.json). This removes the 'settings_file'
key from the readme section on available options to include in the
zappa_settings file to prevent future confusion about 'server side
settings' or the 'settings_file' key in general.